projects/calendar_agent/requirements.txt
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-cov>=4.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pytz>=2024.1 